requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-HTTPAuth
PyMySQL
PyJWT
